# backend/openapi-gen/src/main/java/se/ifmo/gen/model/DragonCreateCoordinates.java
package se.ifmo.gen.model;

import java.net.URI;
import java.util.Objects;
import com.fasterxml.jackson.annotation.JsonProperty;
import com.fasterxml.jackson.annotation.JsonCreator;
import com.fasterxml.jackson.annotation.JsonTypeName;
import org.springframework.lang.Nullable;
import se.ifmo.gen.model.DragonCreateCoordinates;
import org.openapitools.jackson.nullable.JsonNullable;
import java.time.OffsetDateTime;
import jakarta.validation.Valid;
import jakarta.validation.constraints.*;
import io.swagger.v3.oas.annotations.media.Schema;


import java.util.*;
import jakarta.annotation.Generated;

/**
 * Either a DragonCreateCoordinates object or an ID reference to existing coordinates
 */

@Schema(name = "DragonCreate_coordinates", description = "Either a DragonCreateCoordinates object or an ID reference to existing coordinates")
@JsonTypeName("DragonCreate_coordinates")
@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", comments = "Generator version: 7.15.0")
public class DragonCreateCoordinates {

  private @Nullable Integer id;

  private @Nullable Float x;

  private @Nullable Double y;

  public DragonCreateCoordinates id(@Nullable Integer id) {
    this.id = id;
    return this;
  }

  /**
   * Auto-generated unique ID.   Must not be null and must be greater than 0. 
   * minimum: 1
   * @return id
   */
  @Min(1) 
  @Schema(name = "id", description = "Auto-generated unique ID.   Must not be null and must be greater than 0. ", requiredMode = Schema.RequiredMode.NOT_REQUIRED)
  @JsonProperty("id")
  public @Nullable Integer getId() {
    return id;
  }

  public void setId(@Nullable Integer id) {
    this.id = id;
  }

  public DragonCreateCoordinates x(@Nullable Float x) {
    this.x = x;
    return this;
  }

  /**
   * Get x
   * maximum: 187
   * @return x
   */
  @DecimalMax("187") 
  @Schema(name = "x", requiredMode = Schema.RequiredMode.NOT_REQUIRED)
  @JsonProperty("x")
  public @Nullable Float getX() {
    return x;
  }

  public void setX(@Nullable Float x) {
    this.x = x;
  }

  public DragonCreateCoordinates y(@Nullable Double y) {
    this.y = y;
    return this;
  }

  /**
   * Get y
   * @return y
   */
  
  @Schema(name = "y", requiredMode = Schema.RequiredMode.NOT_REQUIRED)
  @JsonProperty("y")
  public @Nullable Double getY() {
    return y;
  }

  public void setY(@Nullable Double y) {
    this.y = y;
  }

  @Override
  public boolean equals(Object o) {
    if (this == o) {
      return true;
    }
    if (o == null || getClass() != o.getClass()) {
      return false;
    }
    DragonCreateCoordinates dragonCreateCoordinates = (DragonCreateCoordinates) o;
    return Objects.equals(this.id, dragonCreateCoordinates.id) &&
        Objects.equals(this.x, dragonCreateCoordinates.x) &&
        Objects.equals(this.y, dragonCreateCoordinates.y);
  }

  @Override
  public int hashCode() {
    return Objects.hash(id, x, y);
  }

  @Override
  public String toString() {
    StringBuilder sb = new StringBuilder();
    sb.append("class DragonCreateCoordinates {\n");
    sb.append("    id: ").append(toIndentedString(id)).append("\n");
    sb.append("    x: ").append(toIndentedString(x)).append("\n");
    sb.append("    y: ").append(toIndentedString(y)).append("\n");
    sb.append("}");
    return sb.toString();
  }

  /**
   * Convert the given object to string with each line indented by 4 spaces
   * (except the first line).
   */
  private String toIndentedString(Object o) {
    if (o == null) {
      return "null";
    }
    return o.toString().replace("\n", "\n    ");
  }
}